# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getData","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1":{"get":{"operationId":"AppController_getV1Data","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/users":{"get":{"operationId":"UserControllerV1_getAllService","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"per_page","required":false,"in":"query","schema":{"type":"number"}},{"name":"sort_by","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort_type","required":false,"in":"query","schema":{"type":"string"}},{"name":"is_deleted","required":false,"in":"query","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/v1/users/{user_id}":{"get":{"operationId":"UserControllerV1_getServiceById","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]},"put":{"operationId":"UserControllerV1_updateUserInfo","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDTO"}}}},"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserControllerV1_deleteService","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/v1/users/changePass/{user_id}":{"put":{"operationId":"UserControllerV1_changePassService","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePassDTO"}}}},"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/v1/classes":{"get":{"operationId":"ClassControllerV1_getAllService","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"per_page","required":false,"in":"query","schema":{"type":"number"}},{"name":"sort_by","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort_type","required":false,"in":"query","schema":{"type":"string"}},{"name":"is_deleted","required":false,"in":"query","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Classes"],"security":[{"bearer":[]}]},"post":{"operationId":"ClassControllerV1_createService","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClassDto"}}}},"responses":{"201":{"description":""}},"tags":["Classes"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}":{"get":{"operationId":"ClassControllerV1_getServiceById","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Classes"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}/role":{"get":{"operationId":"ClassControllerV1_getRoleService","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Classes"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}/people":{"get":{"operationId":"ClassControllerV1_getPeopleServiceById","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Classes"],"security":[{"bearer":[]}]}},"/api/v1/classes/invite":{"post":{"operationId":"ClassControllerV1_inviteService","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InviteUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Classes"],"security":[{"bearer":[]}]}},"/api/v1/classes/invite/accept":{"post":{"operationId":"ClassControllerV1_acceptInviteService","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AcceptInviteUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Classes"],"security":[{"bearer":[]}]}},"/api/v1/classes/join":{"post":{"operationId":"ClassControllerV1_joinService","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserJoinClassDto"}}}},"responses":{"201":{"description":""}},"tags":["Classes"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}/leave":{"put":{"operationId":"ClassControllerV1_leaveService","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Classes"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}/restore":{"patch":{"operationId":"ClassControllerV1_restoreService","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Classes"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}/delete":{"delete":{"operationId":"ClassControllerV1_deleteService","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Classes"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}/remove":{"delete":{"operationId":"ClassControllerV1_removeService","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Classes"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}/students":{"post":{"operationId":"ClassStudentControllerV1_uploadCsv","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"csv":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"tags":["Class Students"]},"get":{"operationId":"ClassStudentControllerV1_getServiceByClassId","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Class Students"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}/students/{student_id}":{"get":{"operationId":"ClassStudentControllerV1_getServiceByStudentId","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"student_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Class Students"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}/students/account-sync":{"put":{"operationId":"ClassStudentControllerV1_syncServiceByStudentId","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountSyncDto"}}}},"responses":{"200":{"description":""}},"tags":["Class Students"],"security":[{"bearer":[]}]}},"/api/v1/import-csvs/upload-csv":{"post":{"operationId":"ImportCsvControllerV1_uploadCsv","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"csv":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"tags":["Import Csvs"]}},"/api/v1/classes/{class_id}/assignments":{"get":{"operationId":"AssignmentControllerV1_getAllService","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"per_page","required":false,"in":"query","schema":{"type":"number"}},{"name":"sort_by","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort_type","required":false,"in":"query","schema":{"type":"string"}},{"name":"is_deleted","required":false,"in":"query","schema":{"type":"string"}},{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Assignments"],"security":[{"bearer":[]}]},"post":{"operationId":"AssignmentControllerV1_createService","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAssignmentDto"}}}},"responses":{"201":{"description":""}},"tags":["Assignments"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}/assignments/{assignment_id}":{"get":{"operationId":"AssignmentControllerV1_getServiceById","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"assignment_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Assignments"],"security":[{"bearer":[]}]},"put":{"operationId":"AssignmentControllerV1_updateService","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"assignment_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignmentInterface"}}}},"responses":{"200":{"description":""}},"tags":["Assignments"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}/assignments/{assignment_id}/restore":{"patch":{"operationId":"AssignmentControllerV1_restoreService","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"assignment_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Assignments"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}/assignments/{assignment_id}/delete":{"delete":{"operationId":"AssignmentControllerV1_deleteService","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"assignment_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Assignments"],"security":[{"bearer":[]}]}},"/api/v1/classes/{class_id}/assignments/{assignment_id}/remove":{"delete":{"operationId":"AssignmentControllerV1_removeService","parameters":[{"name":"class_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"assignment_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Assignments"],"security":[{"bearer":[]}]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Authenticate"],"security":[{"bearer":[]}]}},"/api/v1/auth/register":{"post":{"operationId":"AuthController_createService","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Authenticate"],"security":[{"bearer":[]}]}},"/api/v1/auth/google-activate":{"post":{"operationId":"AuthController_googleCreateService","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginGoogleDto"}}}},"responses":{"201":{"description":""}},"tags":["Authenticate"],"security":[{"bearer":[]}]}},"/api/v1/auth/refresh":{"post":{"operationId":"AuthController_refreshToken","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Authenticate"],"security":[{"bearer":[]}]}},"/api/v1/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Authenticate"],"security":[{"bearer":[]}]}},"/api/v1/auth/logout_all":{"post":{"operationId":"AuthController_logoutAll","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Authenticate"],"security":[{"bearer":[]}]}},"/api/v1/upload-files/image":{"post":{"operationId":"UploadFileControllerV1_uploadImage","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Upload Files"],"security":[{"bearer":[]}]}},"/api/v1/upload-files/{image_name}":{"get":{"operationId":"UploadFileControllerV1_getImageUpload","parameters":[{"name":"XSRF-Token","in":"header","description":"XSRF-Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Upload Files"],"security":[{"bearer":[]}]}}},"info":{"title":"Classroom","description":"The classroom API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ChangePassDTO":{"type":"object","properties":{"old_password":{"type":"string"},"new_password":{"type":"string"},"confirm_password":{"type":"string"}},"required":["old_password","new_password","confirm_password"]},"UpdateUserDTO":{"type":"object","properties":{"first_name":{"type":"string"},"last_name":{"type":"string"},"email":{"type":"string"},"avatar":{"type":"string"}},"required":["first_name","last_name","email","avatar"]},"CreateClassDto":{"type":"object","properties":{"title":{"type":"string"},"section":{"type":"string"},"subject":{"type":"string"},"room":{"type":"string"}},"required":["title","section","subject","room"]},"InviteUserDto":{"type":"object","properties":{"class_id":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","default":"STUDENT"}},"required":["class_id","email","role"]},"AcceptInviteUserDto":{"type":"object","properties":{"class_id":{"type":"string"},"role":{"type":"string","default":"STUDENT"},"code":{"type":"string"}},"required":["class_id","role","code"]},"UserJoinClassDto":{"type":"object","properties":{"role":{"type":"string","default":"STUDENT"},"code":{"type":"string"}},"required":["role","code"]},"AccountSyncDto":{"type":"object","properties":{"student_id":{"type":"string"},"user_id":{"type":"string"}},"required":["student_id","user_id"]},"CreateAssignmentDto":{"type":"object","properties":{"title":{"type":"string"},"instructions":{"type":"string"},"total_points":{"type":"number","default":100},"due_date":{"type":"number"}},"required":["title","instructions","total_points","due_date"]},"AssignmentInterface":{"type":"object","properties":{"_id":{"type":"string"},"updated_at":{"type":"number"},"created_at":{"type":"number"},"deleted_at":{"type":"number"},"class_id":{"type":"string"},"ui_index":{"type":"number"},"title":{"type":"string"},"instructions":{"type":"string"},"total_points":{"type":"number"},"due_date":{"type":"number"}},"required":["_id","updated_at","created_at","deleted_at","class_id","ui_index","title","instructions","total_points","due_date"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"first_name":{"type":"string"},"last_name":{"type":"string"}},"required":["email","password","first_name","last_name"]},"LoginGoogleDto":{"type":"object","properties":{"token_id":{"type":"string"}},"required":["token_id"]}}}}